online_store/requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cssbeautifier==1.14.7
Django==3.2.16
djlint==1.25.0
EditorConfig==0.12.3
gevent==22.10.2
greenlet==2.0.2
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
jsbeautifier==1.14.7
kombu==5.2.4
pathspec==0.11.1
Pillow==8.3.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.3
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.4
stripe==4.0.2
tomli==2.0.1
tqdm==4.65.0
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
zope.event==4.6
zope.interface==6.0
